+++ quill-cassandra-zio/src/test/scala/io/getquill/context/cassandra/zio/examples/other/PlainApp.scala
@@ -23,7 +23,7 @@ object PlainApp {
         .tap(result => printLine(result.toString))
         .provide(zioSession)
 
-    Unsafe.unsafe {
+    Unsafe.unsafe { implicit unsafe =>
       Runtime.default.unsafe.run(czio).getOrThrow()
     }
     ()