+++ quill-sql/src/main/scala/io/getquill/metaprog/Extractors.scala
@@ -14,6 +14,13 @@ class Is[T: Type]:
       None
 
 object Extractors {
+  inline def typeName[T]: String = ${ typeNameImpl[T] }
+  def typeNameImpl[T: Type](using Quotes): Expr[String] =
+    import quotes.reflect._
+    val tpe = TypeRepr.of[T]
+    val name: String = tpe.classSymbol.get.name
+    Expr(name)
+
   def printExpr(using Quotes)(expr: Expr[_], label: String = "") = {
     import quotes.reflect._
     if (label != "")
@@ -405,6 +412,52 @@ object Extractors {
     def unapply(using Quotes)(expr: Expr[Any]): Option[(quotes.reflect.Symbol, quotes.reflect.Term)] =
       import quotes.reflect._
       expr.asTerm.tpe.classSymbol.map(sym => (sym, expr.asTerm))
+
+  object UncastSelectable:
+    def unapply(expr: Expr[_])(using Quotes): Option[Expr[_]] =
+      import quotes.reflect._
+      UncastSelectable.Term.unapply(expr.asTerm).map(_.asExpr)
+    object Term:
+      def unapply(using Quotes)(term: quotes.reflect.Term): Option[quotes.reflect.Term] =
+        import quotes.reflect._
+        term match
+          case AsInstanceOf(inner) => Some(recurse(inner))
+          case other               => Some(other)
+      def recurse(using Quotes)(term: quotes.reflect.Term): quotes.reflect.Term =
+        import quotes.reflect._
+        term match
+          case AsInstanceOf(inner) => recurse(inner)
+          case other               => other
+      private object AsInstanceOf:
+        def unapply(using Quotes)(term: quotes.reflect.Term): Option[quotes.reflect.Term] =
+          import quotes.reflect._
+          term match
+            case TypeApply(Select(inner, "$asInstanceOf$"), _) => Some(inner)
+            case _                                             => None
+  end UncastSelectable
+
+  object Uncast:
+    def unapply(expr: Expr[_])(using Quotes): Option[Expr[_]] =
+      import quotes.reflect._
+      Uncast.Term.unapply(expr.asTerm).map(_.asExpr)
+    object Term:
+      def unapply(using Quotes)(term: quotes.reflect.Term): Option[quotes.reflect.Term] =
+        import quotes.reflect._
+        term match
+          case AsInstanceOf(inner) => Some(recurse(inner))
+          case other               => Some(other)
+      def recurse(using Quotes)(term: quotes.reflect.Term): quotes.reflect.Term =
+        import quotes.reflect._
+        term match
+          case AsInstanceOf(inner) => recurse(inner)
+          case other               => other
+      private object AsInstanceOf:
+        def unapply(using Quotes)(term: quotes.reflect.Term): Option[quotes.reflect.Term] =
+          import quotes.reflect._
+          term match
+            case TypeApply(Select(inner, "asInstanceOf"), _) => Some(inner)
+            case _                                           => None
+  end Uncast
 
   /**
    * Matches `case class Person(first: String, last: String)` creation of the forms: