+++ quill-sql/src/main/scala/io/getquill/parser/Unlifter.scala
@@ -45,12 +45,12 @@ object Unlifter {
     def attempt(expr: Expr[T])(using Quotes): Option[T] =
       import quotes.reflect._
       expr match
-        case '{ SerialHelper.fromSerialized[tt](${ Expr(serial: String) }) } if (TypeRepr.of[tt] <:< TypeRepr.of[Ast]) =>
-          Some(SerialHelper.fromSerialized[Ast](serial).asInstanceOf[T])
+        case '{ SerialHelper.Ast.fromSerialized(${ Expr(serial: String) }).asInstanceOf[t] } =>
+          Some(SerialHelper.Ast.fromSerialized(serial).asInstanceOf[T])
         // On JVM, a Quat must be serialized and then lifted from the serialized state i.e. as a FromSerialized using JVM (due to 64KB method limit)
-        case '{ SerialHelper.QuatProduct.fromSerialized(${ Expr(str: String) }) } =>
+        case '{ SerialHelper.QuatProduct.fromSerialized(${ Expr(str: String) }).asInstanceOf[t] } =>
           Some(SerialHelper.QuatProduct.fromSerialized(str).asInstanceOf[T])
-        case '{ SerialHelper.Quat.fromSerialized(${ Expr(str: String) }) } =>
+        case '{ SerialHelper.Quat.fromSerialized(${ Expr(str: String) }).asInstanceOf[t] } =>
           Some(SerialHelper.Quat.fromSerialized(str).asInstanceOf[T])
         case _ =>
           unlift.lift(expr)
@@ -309,7 +309,7 @@ object Unlifter {
 
   given unliftCaseClass: NiceUnliftable[CaseClass] with
     def unlift =
-      case '{ CaseClass(${ values }: List[(String, Ast)]) } => CaseClass(values.unexpr)
+      case '{ CaseClass(${ name }, ${ values }: List[(String, Ast)]) } => CaseClass(name.unexpr, values.unexpr)
 
   given unliftOperator: NiceUnliftable[Operator] with {
     def unlift =
@@ -350,9 +350,8 @@ object Unlifter {
 
   given quatProductUnliftable: NiceUnliftable[Quat.Product] with {
     def unlift =
-      case '{ Quat.Product.WithRenamesCompact.apply($tpe)(${ Varargs(fields) }: _*)(${ Varargs(values) }: _*)(${ Varargs(renamesFrom) }: _*)(${ Varargs(renamesTo) }: _*) } =>
-        Quat.Product.WithRenamesCompact(tpe.unexpr)(fields.unexprSeq: _*)(values.unexprSeq: _*)(renamesFrom.unexprSeq: _*)(renamesTo.unexprSeq: _*)
-    // case '{ Quat.Product.apply(${Varargs(fields)}: _*) } => Quat.Product(fields.unexprSeq: _*)
+      case '{ Quat.Product.WithRenamesCompact.apply($name, $tpe)(${ Varargs(fields) }: _*)(${ Varargs(values) }: _*)(${ Varargs(renamesFrom) }: _*)(${ Varargs(renamesTo) }: _*) } =>
+        Quat.Product.WithRenamesCompact(name.unexpr, tpe.unexpr)(fields.unexprSeq: _*)(values.unexprSeq: _*)(renamesFrom.unexprSeq: _*)(renamesTo.unexprSeq: _*)
   }
 
   given quatUnliftable: NiceUnliftable[Quat] with {