+++ quill-sql/src/test/scala/io/getquill/quat/QuatSpec.scala
@@ -36,20 +36,20 @@ class QuatSpec extends AnyFreeSpec {
     //   quote(query[TestEnum.MyEnumContainer])
     // }
     "should succeed product-type enum" in {
-      quatOf[TestEnum.ProductEnum] mustEqual Quat.Product("stuff" -> Quat.Value, "otherStuff" -> Quat.Value)
+      quatOf[TestEnum.ProductEnum] mustEqual Quat.Product("ProductEnum", "stuff" -> Quat.Value, "otherStuff" -> Quat.Value)
     }
   }
 
   "boolean and optional boolean" in {
     case class MyPerson(name: String, isHuman: Boolean, isRussian: Option[Boolean])
-    val MyPersonQuat = Quat.Product("name" -> Quat.Value, "isHuman" -> Quat.BooleanValue, "isRussian" -> Quat.BooleanValue)
+    val MyPersonQuat = Quat.Product("MyPersonQuat", "name" -> Quat.Value, "isHuman" -> Quat.BooleanValue, "isRussian" -> Quat.BooleanValue)
     MyPersonQuat mustEqual quatOf[MyPerson]
     quote(query[MyPerson]).ast.quat mustEqual MyPersonQuat
   }
 
   "should support standard case class" in {
     case class MyPerson(firstName: String, lastName: String, age: Int)
-    val MyPersonQuat = Quat.LeafProduct("firstName", "lastName", "age")
+    val MyPersonQuat = Quat.LeafProduct("MyPersonQuat", "firstName", "lastName", "age")
     MyPersonQuat mustEqual quatOf[MyPerson]
     quote(query[MyPerson]).ast.quat mustEqual MyPersonQuat
   }
@@ -57,7 +57,7 @@ class QuatSpec extends AnyFreeSpec {
   "should support embedded" in {
     case class MyName(first: String, last: String) extends Embedded
     case class MyPerson(name: MyName, age: Int) extends Embedded
-    val MyPersonQuat = Quat.Product("name" -> Quat.LeafProduct("first", "last"), "age" -> Quat.Value)
+    val MyPersonQuat = Quat.Product("MyPersonQuat", "name" -> Quat.LeafProduct("MyName", "first", "last"), "age" -> Quat.Value)
     MyPersonQuat mustEqual quatOf[MyPerson]
     quote(query[MyPerson]).ast.quat mustEqual MyPersonQuat
   }
@@ -66,14 +66,14 @@ class QuatSpec extends AnyFreeSpec {
     case class MyName(first: String, last: String) extends Embedded
     case class MyId(name: MyName, memberNum: Int) extends Embedded
     case class MyPerson(name: MyId, age: Int)
-    val MyPersonQuat = Quat.Product("name" -> Quat.Product("name" -> Quat.LeafProduct("first", "last"), "memberNum" -> Quat.Value), "age" -> Quat.Value)
+    val MyPersonQuat = Quat.Product("MyPersonQuat", "name" -> Quat.Product("Prod", "name" -> Quat.LeafProduct("MyName", "first", "last"), "memberNum" -> Quat.Value), "age" -> Quat.Value)
     MyPersonQuat mustEqual quatOf[MyPerson]
     quote(query[MyPerson]).ast.quat mustEqual MyPersonQuat
   }
 
   "should support least upper types" - {
-    val AnimalQuat = Quat.LeafProduct("name")
-    val CatQuat = Quat.LeafProduct("name", "color")
+    val AnimalQuat = Quat.LeafProduct("Animal", "name")
+    val CatQuat = Quat.LeafProduct("Cat", "name", "color")
 
     "simple reduction" in {
       AnimalQuat.leastUpperType(CatQuat).get mustEqual AnimalQuat
@@ -100,9 +100,9 @@ class QuatSpec extends AnyFreeSpec {
   }
 
   "lookup" - {
-    val bar = Quat.Product("baz" -> Quat.Value)
-    val foo = Quat.Product("v" -> Quat.Value, "bar" -> bar)
-    val example = Quat.Product("v" -> Quat.Value, "foo" -> foo)
+    val bar = Quat.Product("bar", "baz" -> Quat.Value)
+    val foo = Quat.Product("foo", "v" -> Quat.Value, "bar" -> bar)
+    val example = Quat.Product("example", "v" -> Quat.Value, "foo" -> foo)
     "path" in {
       example.lookup("foo", true) mustEqual foo
       example.lookup(List("foo", "bar"), true) mustEqual bar
@@ -112,7 +112,7 @@ class QuatSpec extends AnyFreeSpec {
   }
 
   "probit" in {
-    val p: Quat = Quat.Product("foo" -> Quat.Value)
+    val p: Quat = Quat.Product("p", "foo" -> Quat.Value)
     val v: Quat = Quat.Value
     p.probit mustEqual p
     val e = intercept[QuatException] {
@@ -121,8 +121,20 @@ class QuatSpec extends AnyFreeSpec {
   }
 
   "rename" - {
-    val prod = Quat.Product("bv" -> Quat.BooleanValue, "be" -> Quat.BooleanExpression, "v" -> Quat.Value, "p" -> Quat.Product("vv" -> Quat.Value, "pp" -> Quat.Product("ppp" -> Quat.Value)))
-    val expect = Quat.Product("bva" -> Quat.BooleanValue, "be" -> Quat.BooleanExpression, "v" -> Quat.Value, "pa" -> Quat.Product("vv" -> Quat.Value, "pp" -> Quat.Product("ppp" -> Quat.Value)))
+    val prod = Quat.Product(
+      "Prod",
+      "bv" -> Quat.BooleanValue,
+      "be" -> Quat.BooleanExpression,
+      "v" -> Quat.Value,
+      "p" -> Quat.Product("Emb", "vv" -> Quat.Value, "pp" -> Quat.Product("EmbSingle", "ppp" -> Quat.Value))
+    )
+    val expect = Quat.Product(
+      "expect",
+      "bva" -> Quat.BooleanValue,
+      "be" -> Quat.BooleanExpression,
+      "v" -> Quat.Value,
+      "pa" -> Quat.Product("Emb", "vv" -> Quat.Value, "pp" -> Quat.Product("EmbSingle", "ppp" -> Quat.Value))
+    )
     val value = Quat.Value
     "rename field" in {
       prod.withRenames(List("bv" -> "bva", "p" -> "pa")).applyRenames mustEqual expect
@@ -135,7 +147,7 @@ class QuatSpec extends AnyFreeSpec {
   "should serialize" - {
     // Need to import implicits from BooQuatSerializer otherwise c_jl_UnsupportedOperationException happens in JS
     import BooQuatSerializer._
-    val example = Quat.Product("bv" -> Quat.BooleanValue, "be" -> Quat.BooleanExpression, "v" -> Quat.Value, "p" -> Quat.Product("vv" -> Quat.Value))
+    val example = Quat.Product("Prod", "bv" -> Quat.BooleanValue, "be" -> Quat.BooleanExpression, "v" -> Quat.Value, "p" -> Quat.Product("EmbSingle", "vv" -> Quat.Value))
     "with boo" in {
       Quat.fromSerialized(serialize(example)) mustEqual example
     }
@@ -188,7 +200,7 @@ class QuatSpec extends AnyFreeSpec {
       def func = quote {
         (q: Query[MyPerson]) => q.filter(p => p.name == "Joe")
       }
-      func.ast.quat mustEqual Quat.Product("name" -> Quat.Value, "isRussian" -> Quat.BooleanValue)
+      func.ast.quat mustEqual Quat.Product("Prod", "name" -> Quat.Value, "isRussian" -> Quat.BooleanValue)
     }
     "case class with boundary" in {
       case class MyPerson(name: String, isRussian: Boolean)
@@ -274,7 +286,7 @@ class QuatSpec extends AnyFreeSpec {
       // def func[T <: { def name: String; def isRussian: Boolean }] = quote {
       //   (q: T) => q
       // }
-      // func.ast.quat mustEqual Quat.Product("name" -> Quat.Value, "isRussian" -> Quat.BooleanValue).withType(Quat.Product.Type.Abstract)
+      // func.ast.quat mustEqual Quat.Product("Prod", "name" -> Quat.Value, "isRussian" -> Quat.BooleanValue).withType(Quat.Product.Type.Abstract)
     }
     "structural with bool indirect" in {
       // TODO Figure out issues with structural types
@@ -283,14 +295,14 @@ class QuatSpec extends AnyFreeSpec {
       // def func[T <: { def name: String; def isRussian: Bool }] = quote {
       //   (q: T) => q
       // }
-      // func.ast.quat mustEqual Quat.Product("name" -> Quat.Value, "isRussian" -> Quat.BooleanValue)
+      // func.ast.quat mustEqual Quat.Product("Prod", "name" -> Quat.Value, "isRussian" -> Quat.BooleanValue)
     }
     "case class" in {
       case class MyPerson(name: String, isRussian: Boolean)
       def func = quote {
         (q: MyPerson) => q
       }
-      func.ast.quat mustEqual Quat.Product("name" -> Quat.Value, "isRussian" -> Quat.BooleanValue)
+      func.ast.quat mustEqual Quat.Product("Prod", "name" -> Quat.Value, "isRussian" -> Quat.BooleanValue)
     }
     "case class with boundary" in {
       case class MyPerson(name: String, isRussian: Boolean)
@@ -298,7 +310,7 @@ class QuatSpec extends AnyFreeSpec {
       // def func[T <: MyPerson] = quote {
       //   (q: T) => q
       // }
-      // func.ast.quat mustEqual Quat.Product("name" -> Quat.Value, "isRussian" -> Quat.BooleanValue)
+      // func.ast.quat mustEqual Quat.Product("Prod", "name" -> Quat.Value, "isRussian" -> Quat.BooleanValue)
     }
     "interface" in {
       trait LikePerson { def name: String; def isRussian: Boolean }
@@ -313,7 +325,7 @@ class QuatSpec extends AnyFreeSpec {
       // def func[T <: LikePerson] = quote {
       //   (q: T) => q
       // }
-      // func.ast.quat mustEqual Quat.Product("name" -> Quat.Value, "isRussian" -> Quat.BooleanValue)
+      // func.ast.quat mustEqual Quat.Product("Prod", "name" -> Quat.Value, "isRussian" -> Quat.BooleanValue)
     }
     "interface with boundary boolean indirect" in {
       type Bool = Boolean
@@ -322,7 +334,7 @@ class QuatSpec extends AnyFreeSpec {
       // def func[T <: LikePerson] = quote {
       //   (q: T) => q
       // }
-      // func.ast.quat mustEqual Quat.Product("name" -> Quat.Value, "isRussian" -> Quat.BooleanValue)
+      // func.ast.quat mustEqual Quat.Product("Prod", "name" -> Quat.Value, "isRussian" -> Quat.BooleanValue)
     }
     "boundary with value" in {
       def func[T <: Int] = quote {