+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# 4.6.0.1
+
+- [Upgrade to latest version of ZIO (2.0.8)](https://github.com/zio/zio-protoquill/pull/242)
+- [JDBC DataSource should not need to be closeable](https://github.com/zio/zio-protoquill/pull/205)
+
+# 4.6.0
+
+- [Json/Jsonb Encoding for Postgres. Bump ZIO](https://github.com/zio/zio-protoquill/pull/202)
+
+# 4.5.0
+
+- [Implement Dynamic Query DSL for ProtoQuill](https://github.com/zio/zio-protoquill/pull/198)
+- [Fix dynamic Insert/UpdateMeta and better warning message](https://github.com/zio/zio-protoquill/pull/197)
+- [Support for structural types in Quoted functions](https://github.com/zio/zio-protoquill/pull/199)
+- [Upstream changes. Add JDBC encoders for java.time dates](https://github.com/zio/zio-protoquill/pull/191)
+
 # 4.4.1
 
 - [Fixing Caliban Datatype conversion issues](https://github.com/zio/zio-protoquill/pull/188)